requirements.txt
flask
psycopg2-binary # PostgreSQL adapter
python-dotenv
pytest
pytest-flask
pytest-cov # test cov report
